% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ability.R
\name{ability}
\alias{ability}
\title{Ability}
\usage{
ability(rank, satPercent, score, nodePosition)
}
\arguments{
\item{R}{This is the Rank of the maze.}

\item{Score}{This is the series of connected paths via the black dots}
}
\description{
Ability
}
\details{
The ability function returns the weighted score of the individual given his raw score (i.e. the number of black dotes collected).
}
\author{
Aiden Loe and Maria Sanchez
}

