% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIG.R
\name{AIG}
\alias{AIG}
\title{AIG: A package for studying automatic item generators.}
\description{
The AIG package contains methods for studying automatic item generators. The package includes several item types generations for research purposes. If the generator has many related functions, then it will be pushed into a new package instead. Other generators include, netGen, mazeGen, numGen.
}
\details{
Still need to make changes here.
}
\section{Linear Syllogism Generation}{

This function creates linear syllogsim tasks accordingly to your specifications. \cr
\code{\link{lisy}}
}

