% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/blackNodeRoutes.R
\name{blackNodeRoutes}
\alias{blackNodeRoutes}
\title{blackNodeRoutes}
\usage{
blackNodeRoutes(rank, nodePosition)
}
\arguments{
\item{nodePosition}{tells you all the position of the black dots}

\item{R}{This is the Rank of the maze.}
}
\description{
blackNodeRoutes
}
\details{
This function returns the frequently of paths going through a sequence of black dots
}
\author{
Aiden Loe and Maria
}

