% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colleage.R
\name{colleage}
\alias{colleage}
\title{colleage}
\usage{
colleage(n, R)
}
\arguments{
\item{n}{This is the node number.}

\item{R}{This is the Rank of the maze.}
}
\description{
colleage
}
\details{
This function tells us which nodes are connected to the node you are interested in based on a given Grid.
}
\author{
Aiden Loe
}

