% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_colours.R
\name{colour_display}
\alias{colour_display}
\title{Selecting the colour display of the nodes}
\usage{
colour_display(base.colour, start.colour, end.colour)
}
\description{
Selecting the colour display of the nodes
}
\author{
Aiden Loe
}

