% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colourNodePosition.R
\name{colourNodePosition}
\alias{colourNodePosition}
\title{colourNodePosition}
\usage{
colourNodePosition(rank, satPercent, seed)
}
\arguments{
\item{satPercent}{Percentage of saturation.}

\item{Rank}{This is the Rank of the maze.}

\item{setseed.}{To always get the same position.}
}
\description{
This function will not sample from the first node position. If you consider sampling from the first node, then in javascript, the summing of the black dotes need to begin from 1 rather than 0. To keep it simple, always ensure that the first node is not sampled as a black dot.
}
\details{
Returns the colour node position. You need to use the node position function first.
}
\examples{
\dontrun{

colourNodePosition(rank=3,satPercent=0.5,seed=1)

}
}
\author{
Aiden Loe
}

