% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/difficulty.R
\name{diff}
\alias{diff}
\title{Maze Diffculty}
\usage{
diff(rank, satPercent, nodePosition)
}
\arguments{
\item{satPercent}{This is the saturation}

\item{nodePosition}{This is the distribution of the colour node positions}

\item{R}{This is the Rank of the maze.}
}
\description{
Maze Diffculty
}
\details{
This function tells us the difficulty level of the rank given a saturation and black node distribution
}
\examples{
\dontrun{

rank <- 5
satPercent <- 0.5

#Black nodes distribution
nodePosition <- colourNodePosition(rank,satPercent,1)

#calculate difficulty
diff(rank,satPercent ,nodePosition)
}
}
\author{
Aiden Loe and Maria Sanchez
}

