% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edge_weights.R
\name{edge.v}
\alias{edge.v}
\title{edge.v}
\usage{
edge.v(edge.value = NULL)
}
\description{
This allows us to select the weights of the edges
}
\details{
This tells us that the value of the edge is dependent on the randomly drawing
number 1:10
}
\author{
Aiden Loe
}

