% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logic_maps.R
\name{g.map}
\alias{g.map}
\title{g.map}
\usage{
g.map(x, base.colour, start.colour, end.colour, z = NULL, newValue,
  default.colour = TRUE, edge.value = 0, no.label = TRUE)
}
\arguments{
\item{x}{This is taken from the logic derived using the igraph package.}

\item{base.colour}{This is the colour of all the nodes if no colour is specified}

\item{start.colour}{This is the colour of the first node}

\item{end.colour}{This is the colour of the last node}

\item{z}{If Z=NUll then use default names in the package}

\item{newValue}{This is the value of the number of nodes from the logic}

\item{edge.value}{This depends on the width of the edge}
}
\value{
The map item based on the logic of the igraph package
}
\description{
Map items that that does require a check
}
\details{
This functions is embedded with \code{check.graph}, \code{edge.v}, \code{colour_display}
}
\author{
Aiden Loe
}

