% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genmaze.R
\name{genMaze}
\alias{genMaze}
\title{genMaze}
\usage{
genMaze(R)
}
\arguments{
\item{R}{This is the Rank of the maze.}
}
\description{
genMaze
}
\details{
The Genmaze function generates the list of edges
}
\author{
Aiden Loe
}

