% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genuniquesolution.R
\name{genUniqueSolution}
\alias{genUniqueSolution}
\title{genUniqueSolution}
\usage{
genUniqueSolution(rank, satPercent, seed)
}
\arguments{
\item{R}{This is the Rank of the maze.}
}
\description{
genUniqueSolution
}
\details{
The generate unique solution function searches for the SEED that returns only one unique solution for a given rank and saturation.
}
\examples{
\dontrun{

rank <- 5
satPercent <- 0.5

#Number of unique solutions
lookUniqueSolution(rank,satPercent,11)

#Searches for just one unique solution
justOne <- genUniqueSolution(rank,satPercent,11)

}
}
\author{
Aiden Loe and Maria Sanchez
}

