% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grid.R
\docType{data}
\name{gridFour}
\alias{gridFour}
\title{Four Grid}
\format{A data frame with 2 columns
\describe{
  \item{start}{Coordinates of Start Node.}
  \item{end}{Coordinates End Node.}
}}
\usage{
data(gridFour)
}
\description{
This returns a four grid maze. This are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 4
data(gridFour)
coordinates <- gridFour

}
}
\keyword{datasets}

