% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grid.R
\docType{data}
\name{gridThree}
\alias{gridThree}
\title{Three Grid}
\format{A data frame with 2 columns
\describe{
  \item{start}{start, coordinates of Start Node.}
  \item{carat}{end, coordinates End Node.}
}}
\usage{
data(gridThree)
}
\description{
This returns a three grid maze. This are standardized coordinates.
}
\examples{
\dontrun{

# Returns a Grid with rank = 3
data(gridThree)
coordinates <- gridThree

}
}
\keyword{datasets}

