% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/howmany.R
\name{howMany}
\alias{howMany}
\title{howMany}
\usage{
howMany(rank, satPercent)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{satPercent}{Percentage of saturation.}
}
\description{
howMany
}
\details{
Calculates how many possible variation of black dotes for a given saturation.
Does not account for the first node being a black dot.
}
\author{
Aiden Loe
}

