% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARMadd.R
\name{ICARarith}
\alias{ICARarith}
\title{ICAR Arithmetic Generator}
\usage{
ICARarith(numProbs = 5, numOps = 1, magnitude = 1, arithOps = "add",
  allowNeg = FALSE, sameValue = FALSE, order = "mixed")
}
\arguments{
\item{numProbs}{Number of desired problems}

\item{numOps}{sets the number of operations in the stimuli (range of values is 1 to 5)}

\item{magnitude}{sets the range of values for the integers in the stimuli by powers of 10 (range of values is 1 to 5)}

\item{arithOps}{There are two options in this argument, either 'add' or 'subtract', "multiply.}

\item{allowNeg}{Allow the use of negative values. Only for "add" arithmetic operator. Not allowed for "subtract".}

\item{sameValue}{Deciding if you want the values to be the same or different values.}

\item{order}{There are three options to order the values presented. "mixed", "decreasing", "increasing". Does not matter when same Value = TRUE.}
}
\value{
\describe{
\item{Question}{The arithmetic questions that are automatically generated.}
\item{Response Options}{Eight response options. }
\item{Answers}{The answer to the question generated.}
\item{Code}{First digit is the magnitude. Subsequent digits correspond to the positional values in the question. 0 is positive values, 1 is negative values.}
    }
}
\description{
A basic arithmetic item generator for ICAR.
}
\details{
The arithmetic function can currently generate additive and substractive operators. For the additive operators, you have the option of using both negative and positive numerical values. For the substractive operators, you are currently restricted to only using positive numerical values.
}
\examples{

ICARarith(numProbs=4, numOps = 4, magnitude = 3, arithOps = "subtract",
 allowNeg = FALSE, sameValue=FALSE, order = "mixed")


}
\author{
Aiden Loe and David Condon
}

