% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lookuniquesolution.R
\name{lookUniqueSolution}
\alias{lookUniqueSolution}
\title{lookUniqueSolution}
\usage{
lookUniqueSolution(rank, satPercent, seed)
}
\arguments{
\item{R}{This is the Rank of the maze.}
}
\description{
lookUniqueSolution
}
\details{
The unique solution function returns the number of unique series of black dots for a given rank, saturation and seed.
}
\examples{
\dontrun{

rank <- 5
satPercent <- 0.5

#Number of unique solutions
lookUniqueSolution(rank,satPercent,1)

}
}
\author{
Aiden Loe and Maria Sanchez
}

