% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxScoreRoutes.R
\name{maxScoreRoutes}
\alias{maxScoreRoutes}
\title{maxScoreRoutes}
\usage{
maxScoreRoutes(rank, nodePosition)
}
\arguments{
\item{R}{This is the Rank of the maze.}
}
\description{
maxScoreRoutes
}
\details{
The maxScoreRoutes function tells you the possible routes to achieve a maximum score based on the colourNode position.
You need to use the nodePosition function first prior to using this.
}
\author{
Aiden Loe and Maria
}

