% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maze.R
\name{maze}
\alias{maze}
\title{maze}
\usage{
maze(rank, i, saturation, grid, html)
}
\arguments{
\item{rank}{This is the Rank of the maze.}

\item{i}{To make sure that the randomness of the created black dots is captured and not repeated.}

\item{saturation}{The number of black dots created for a given grid.}

\item{html}{is the folder to save the source code in}

\item{Grid}{is the grid of the maze}
}
\description{
maze
}
\details{
This function creates a maze into your folder.
A grid object needs to be called out first before runing the maze function.
The grid object needs to be the same as the rank given.
}
\examples{
\dontrun{

rank <- 3
i <- 2
saturation <- 0.5

# Grid same as rank
data(gridThree)

# Folder to save html/css/java script
filePath = getwd()
maze(rank,i,saturation,gridThree,filePath)

}
}
\author{
Aiden Loe
}

