% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minLegRoutes.R
\name{minLegRoutes}
\alias{minLegRoutes}
\title{minLegRoutes}
\usage{
minLegRoutes(rank, nodePosition)
}
\arguments{
\item{rank}{This is the Rank of the maze.}
}
\description{
minLegRoutes
}
\details{
The minLegRoutes function tells you the possible routes to achieve a maximum score based on the colourNode position with a minimum number legs.
You need to use the nodePosition function first prior to using this.
}
\author{
Aiden Loe and Maria
}

