% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minStep.R
\name{minStep}
\alias{minStep}
\title{minStep}
\usage{
minStep(rank, nodePosition)
}
\arguments{
\item{Rank}{This is the Rank of the maze.}

\item{minStep}{This returns the minimum step achievable for a given max score}
}
\description{
minStep
}
\details{
This function tells us the minimum steps to achieve the maximums core given the colour node positions.
}
\examples{
\dontrun{

rank <- 5
satPercent <- 0.5

nodeLength(rank)# longest row

nodePosition <- colourNodePosition(rank, satPercent, 2) ### always set seed.

#Get Max Score
maxScore(rank, nodePosition)

maxRoutes <- maxScoreRoutes(rank, nodePosition)

# Get minimum steps required to get maximum score
minStep(rank, nodePosition)
}
}
\author{
Aiden Loe
}

