% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_1.R
\name{nMAdd}
\alias{nMAdd}
\title{nMAdd}
\usage{
nMAdd(length, items)
}
\arguments{
\item{length}{This is the length of each item}

\item{items}{The number of items you want to generate}
}
\description{
nMAdd
}
\details{
This is based on linearity rule / pattern recognition rule
}
\examples{
\dontrun{

nMAdd(4,5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

