% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_3.R
\name{nmCombo}
\alias{nmCombo}
\title{nmCombo}
\usage{
nmCombo(items, seed)
}
\arguments{
\item{items}{Number of items you want to generate.}

\item{seed}{This gives you the same result again.}
}
\description{
i.e combination of:
}
\details{
2 linear sequences (without use of arithmetic)
}
\examples{
\dontrun{

nmCombo(items=5, seed=5)

}
}
\author{
Aiden Loe and Filip Simonfy
}

