% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel1_1.R
\name{nMOne}
\alias{nMOne}
\title{nMOne}
\usage{
nMOne(x, items)
}
\arguments{
\item{x}{To get the length value}

\item{Items}{The number of items you have to generate.}
}
\description{
nMOne
}
\details{
counting the alphabets representation.
}
\examples{
\dontrun{

nMOne(3,26)

}
}
\author{
Aiden Loe and Filip Simonfy
}

