% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel3_1.R
\name{nmParaSeq}
\alias{nmParaSeq}
\title{nmParaSeq}
\usage{
nmParaSeq(add, multi, sub, div, items)
}
\arguments{
\item{add}{Generate the number of items that uses the addition operator.}

\item{multi}{Generate the number of items that uses the multiplication operator.}

\item{sub}{Generate the number of items that uses the substraction operator.}

\item{div}{Generate the number of items that uses the division operator.}

\item{items}{Generate a random mix of items.}
}
\description{
i.e combination of:
}
\details{
2 linear sequences (With use of arithmetic operator)
}
\examples{
\dontrun{

nmParaSeq(add=2,multi=2,sub=2,div=4,items=10)

}
}
\author{
Aiden Loe and Filip Simonfy
}

