% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_2.R
\name{nMRepresent}
\alias{nMRepresent}
\title{nMRepresent}
\usage{
nMRepresent(numLetter, value, items, random = FALSE)
}
\arguments{
\item{numLetter}{Number of Letters you want.}

\item{value}{Value + numLetter gives you back the total number of letters.}

\item{items}{Number of items you want to generate.}
}
\description{
This function tells you the count of a specific letter.
}
\details{
Returns the number representation which is the count of the specific letters.
}
\examples{
\dontrun{

nMRepresent(numLetter= 2 ,value=3, items=10, random=FALSE)

}
}
\author{
Aiden Loe
}

