% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_1.R
\name{nMThree}
\alias{nMThree}
\title{nMThree}
\usage{
nMThree(items, n, arithmetic)
}
\arguments{
\item{items}{The number of items you want to generate}

\item{n}{Value you want use the arithmetic operator on}

\item{arithmetic}{Use either 'add', 'substr', 'multi', 'div'.}
}
\description{
nMThree
}
\details{
This is based on linearity rule / pattern recognition rule and also arimetic abilty. Currently it only displays up to a series of 9.
}
\examples{
\dontrun{

nMThree(items=4,n=2,arithmetic="add")

}
}
\author{
Aiden Loe and Filip Simonfy
}

