% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel2_1.R
\name{nMThreeAlpha}
\alias{nMThreeAlpha}
\title{nMThreeAlpha}
\usage{
nMThreeAlpha(items, reverse = FALSE)
}
\arguments{
\item{items}{The number of items you want to generate}

\item{reverse}{Changing reverse to TRUE will flip to direction of the progressive alphabets}
}
\description{
nMThreeAlpha
}
\details{
This is based on linearity rule / pattern recognition rule
This is for alphabets and is restricted to 21 items.
}
\examples{
\dontrun{

nMThreeAlpha(3,reverse=TRUE)

}
}
\author{
Aiden Loe and Filip Simonfy
}

