% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel1_2.R
\name{nMTwo}
\alias{nMTwo}
\title{nMTwo}
\usage{
nMTwo(x, items, random = FALSE)
}
\arguments{
\item{x}{To get the length value}

\item{random}{The alphabets are not in order.}

\item{Items}{The number of items you have to generate.}
}
\description{
This function will only return two letter groups of different length.
}
\details{
counting the alphabets representation.
}
\examples{
\dontrun{

nMTwo(3,6, random=FALSE)
}
}
\author{
Aiden Loe and Filip Simonfy
}

