% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nModel1_3.R
\name{numModelCat}
\alias{numModelCat}
\title{numModelCat}
\usage{
numModelCat(cat, items, seed, random = FALSE)
}
\arguments{
\item{cat}{Number of categorical groups per question.}

\item{items}{The number of items you want to generate.}

\item{seed}{To get the same values again.}

\item{random}{To randomise the position of the numeric values.}
}
\description{
numModelCat
}
\details{
This is based on the categorical / pattern recognition rule
}
\examples{
\dontrun{

numModelCat(2,4,1,random=TRUE)

}
}
\author{
Aiden Loe and Filip Simonfy
}

