% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMap.R
\name{plot.multi.map}
\alias{plot.multi.map}
\title{plot.multi.map
#'}
\usage{
\method{plot}{multi.map}(x, n, v.size, height, width, png = TRUE, layout,
  vertex.label.cex, vertex.shape, xlab, ylab, cex.lab)
}
\arguments{
\item{x}{is the object created by \code{closedMaps} or \code{g.maps}}

\item{n}{is the number of times you want to generate the plot}

\item{v.size}{Is the size of the vertice}

\item{height}{This is the height of the image}

\item{width}{This is the width of the image}

\item{png}{Tells you to print to your working directory automatically or not.}

\item{layout}{Layout of the graph which graph}

\item{vertex.label.cex}{This is the size of the vertex label}

\item{vertex.shape}{This determines the shape of the vertex shape}

\item{xlab}{The label of the x coordinates}

\item{ylab}{The label of the y coordinates}

\item{cex.lab}{The size of the labels}

\item{seed}{This will always result in the same graph}

\item{png_name}{This is the input of the name}
}
\description{
plot.multi.map
#'
}

