% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solution.R
\name{solution}
\alias{solution}
\title{solution}
\usage{
solution(rank, nodePosition)
}
\arguments{
\item{nodePosition}{tells you all the position of the black dots}

\item{R}{This is the Rank of the maze}

\item{set.seed}{so that the same random selected black dots will ables be the same.}
}
\description{
solution
}
\details{
This function tells us the solution for all series of routes including both black and non black dots.
}
\author{
Aiden Loe and Maria Sanchez
}

