% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR3d.R
\name{spatial3d}
\alias{spatial3d}
\title{Spatial 3D Reasoning Item}
\usage{
spatial3d(seed = 1, theta = 380, phi = 50)
}
\arguments{
\item{seed}{To generate the same set of item(s) on the local computer.}

\item{theta}{Changes the polar coordinates}

\item{phi}{Changes the polar coordinates}
}
\value{
\describe{
\item{figure}{Return the matrix that generates the display figure.}
    }
}
\description{
This function generates a 3 dimensional display figure. It acts as a wrap because the creation of the figure is done using functions from the rgl package.
}
\details{
For 3D figures, some of the cubes may be hidden in sight when automatically generated. Hence, one would need to rotate the display figure several times to ensure that none of the cubes are hidden. The rotation can be done by changing the theta and phi value in the function. To ensure that the same image is generated again, please provide a seed value. You will need to use the rgl post script function to save the figure. Currently, the function returns the matrix that is used to generate the display figure. The matrix can be used for the \code{\link{spatial3d_mirror}} to generate a mirror image of the display figure.

Apart from changing the polar coordiates (theta and phi), you can also rotate the figure manually by clicking on the figure and moving it in different direction.
}
\examples{
a <- spatial3d(seed=4)

# To save the figure (not run)
# wd<- '~/desktop'
# item <- 1
# save <- paste0(wd,"/display3d_",item,".pdf")
# rgl.postscript(save,"pdf")
}
\author{
Aiden Loe
}
\seealso{
\code{\link{lisy}}, \code{\link{arith}}, \code{\link{spatial2d}}, \code{\link{spatial3d_mirror}}
}

