% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MR3d.R
\name{spatial3d_mirror}
\alias{spatial3d_mirror}
\title{Mirror Spatial 3D Reasoning Item}
\usage{
spatial3d_mirror(x)
}
\arguments{
\item{x}{An object with class of threeD.}
}
\value{
\describe{
\item{figure}{Return the matrix that generates the mirror image of the display figure.}
    }
}
\description{
This function generates the mirror image of the 3 dimensional display figure. It acts as a wrap because the creation of the figure is done using functions from the rgl package.
}
\details{
For 3D figures, some of the cubes may be hidden in sight when automatically generated. Hence, one would need to rotate the display figure several times to ensure that none of the cubes are hidden. You can also rotate the figure manually by clicking on the figure and moving it in different direction.
}
\examples{
display <- spatial3d(seed=4)
display_mirror <- spatial3d_mirror(display)

#' # To save the figure (not run)
# wd<- '~/desktop'
# item <- 1
# save <- paste0(wd,"/mirror3d_",item,".pdf")
# rgl.postscript(save,"pdf")
}
\author{
Aiden Loe
}
\seealso{
\code{\link{lisy}}, \code{\link{arith}}, \code{\link{spatial2d}}, \code{\link{spatial3d}}
}

