% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRtwoD.R
\name{twoD}
\alias{twoD}
\title{2D Reasoning Items}
\usage{
twoD(items, wd = NULL, degree = 360)
}
\arguments{
\item{items}{The number of items to generate.}

\item{wd}{This is the working directory to save the figures in. If not provided, the file will be saved in the current working directory.}

\item{degree}{This allows you to change the rotation of the display figure by increasing or decreasing the value.}
}
\value{
\describe{
\item{ans}{Return the matrix that generates the display figure and answer.}
\item{mirror}{Return the matrix that generates the mirror figure.}
\item{dist3}{Return the matrix that generates the distractor without the first square.}
\item{dist4}{Return the matrix that generates the distractor that addes an extra square.}
    }
}
\description{
This function generates a 2 dimensional figure with 4 distractors and 1 answer that is
}
\details{
By default the actual answer is always at half of the display polar coordinates. The display figure polar coordinates by default is 360. Increasing the degrees will make the figure rotate clock-wise and decreasing the degrees will make the figure rotate anti-clockwise.

 4 distractors are generated. 2 of the generated distractors are a mirror image of the displayed figure. The difference in the polar coordinates between the two distractors are fixed at 120. The third distractor is a figure wihout the first square that is generated by the display item. The fourth distractor is a figure with an additional square added to the distractor. Currently, the polar coordinates are fixed for the third and fourth distractor as well.
}
\examples{
wd<- "~/desktop"
twoD(items=1, wd=wd, degree=180)
}
\author{
Aiden Loe
}
\seealso{
\code{\link{lisy}}, \code{\link{arith}}
}

