% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIG.R
\name{AIG}
\alias{AIG}
\title{AIG: A package for studying Automatic Item Generators.}
\description{
The AIG package contains generators for studying Automatic Item Generators.
}
\details{
The package includes several item types generations for research purposes. Currently, it can only generate linear syllogism items.
}
\section{Linear Syllogism Generation}{

This function creates linear syllogsim items according to your specifications. \cr
\code{\link{lisy}}
}

