% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIG.R
\name{AIG}
\alias{AIG}
\title{AIG: A package for studying Automatic Item Generators.}
\description{
The AIG package contains generators for studying automatically generated item types.
}
\details{
The package will include several generators for psychological research purposes. Currently, it can only generate linear syllogism items.
}
\section{Linear Syllogism Generator}{

This function creates linear syllogsim items according to certain specifications given by the arguments. \cr
\code{\link{lisy}}
}

\section{Arithmetic Generator}{

This function creates arithmetic items according to certain specifications given by the arguments. \cr
\code{\link{arith}}
}
\references{
Sternberg, R. J. (1980). Representation and process in linear syllogistic reasoning. Journal of Experimental Psychology: General, 109(2), 119.

Embretson, S. E. (1999). Generating items during testing: Psychometric issues and models. Psychometrika, 64(4), 407-433.
}

