% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lisy.R
\name{lisy}
\alias{lisy}
\title{lisy}
\usage{
lisy(seed = 1, nclues = 4, nspread = 5, reverseprob = 0.5, Ndist = 4,
  incidentals = "names", dist = "mixed", itemSet = "random",
  items = NULL, scales = NULL)
}
\arguments{
\item{seed}{This generates the same question again from local computer.}

\item{nclues}{This is the number of sentences to make up the item}

\item{nspread}{This is the spread of the number of incidentals in total.}

\item{reverseprob}{This calculates how you want to divide the comparison terms.}

\item{Ndist}{This returns the number of distractors per question.}

\item{incidentals}{This tells the function whether the item features are 'names' or 'objects'.}

\item{dist}{This allows you to select the type of distractors. The number of false distractors must be less than the number of clues.}

\item{itemSet}{This is the choice of itemset you want. If itemset='random' then the generator will randomly select one (People, Fruits, Superheroes). Change itemset='own' if you are using your own item set.}

\item{items}{This inputs your own item type. At least 10 items.}

\item{scales}{This is the comparison terms. At least 2 comparison terms (i.e."bigger","smaller")}
}
\description{
This function generates linear syllogistic reasoning items.
}
\details{
There are several things to bear in mind. To use own item set, please have at least 10 items within the itemset. In order for scale comparison to make sure. Please ensure that you have at least 2 comparisons. The function will stop if the criteria is not met. The genearation of items are slower if you have a huge item set.

When nspread and nclue is = 3. This means that there are 3 sentences, and only 3 names. This makes it impossible to generate an invalid distractor. As such, only the false distractors will be created. Since there are only three clues, then 3 false distractors will be created.

When nspread and nclues are the same. All the names of the invalid distractors will be taken from the names that are used in the clues. As nspread value increases, the likelihood of having names not taken from the clues increases.
}
\examples{
\dontrun{


 # Generate an item
 item <-  lisy(seed=4, nclues=5, nspread=5,reverseprob=.5,
 Ndist=4, incidentals='names', dist="false",
 itemSet='random',items= NULL, scales = NULL)

 # Generate 100 items
 nitems <- 100

 # Use different number of names and clues
 params <- data.frame(seed=1:nitems,
 nclues=ceiling((1:nitems)/20)+3,
 nspread=ceiling((1:nitems)/15)+4)

 # Loop through
 qtable <- NULL
 for (i in 1:nitems) {
 runs <- lisy(seed=i,
 nclues=params$nclues[i],
 nspread=params$nspread[i],
 reverseprob=.5,  Ndist=4, incidentals='names',
 dist="mixed",itemSet='random',items= NULL, scales = NULL)
 qtable[[i]] <- runs
 }
 qtable

 # Save csv file
 write.csv(do.call("rbind",qtable), file="~/desktop/test.csv"  )

 #############
 # Example using own item set
 library("babynames")
 bNames <- sapply(babynames[,3], as.character)
 compare <- c("taller", "older", "smaller", "bigger","stronger", "weaker")

 #Generate items
 lisy(seed=4, nclues=5, nspread=7,reverseprob=.5, Ndist=2,
 incidentals= 'names',dist="false",
 itemSet='own',items= bNames, scales = compare)

 item
}



}
\author{
Aiden Loe and Francis Smart
}

